requirements/base.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.13.0
django-debug-toolbar==4.0.0
django-filter==23.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2023.3
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
